Code cleanup & refactored targeting enums
