+++ Orrery Combat Framework - Heart Module/Data/Scripts/HeartModule/Definitions/StandardClasses/DefinitionContainer.cs
@@ -1,11 +1,6 @@
 ﻿using Heart_Module.Data.Scripts.HeartModule.Projectiles.StandardClasses;
 using Heart_Module.Data.Scripts.HeartModule.Weapons.StandardClasses;
 using ProtoBuf;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 
 namespace Heart_Module.Data.Scripts.HeartModule.Definitions.StandardClasses
 {
