+++ Orrery Combat Framework - Heart Module/Data/Scripts/HeartModule/HeartData.cs
@@ -5,7 +5,6 @@ using Sandbox.ModAPI;
 using System;
 using System.Collections.Generic;
 using VRage.Game.ModAPI;
-using VRage.ModAPI;
 
 namespace Heart_Module.Data.Scripts.HeartModule
 {
