+++ Orrery Combat Framework - Heart Module/Data/Scripts/HeartModule/Projectiles/GuidanceHelpers/ProjectileGuidance.cs
@@ -1,5 +1,4 @@
-﻿using Heart_Module.Data.Scripts.HeartModule.Debug;
-using Heart_Module.Data.Scripts.HeartModule.Projectiles.StandardClasses;
+﻿using Heart_Module.Data.Scripts.HeartModule.Projectiles.StandardClasses;
 using Heart_Module.Data.Scripts.HeartModule.Utility;
 using Sandbox.ModAPI;
 using System.Collections.Generic;
