+++ Orrery Combat Framework - Heart Module/Data/Scripts/HeartModule/Projectiles/ProjectileDefinitionManager.cs
@@ -1,8 +1,5 @@
-﻿using Heart_Module.Data.Scripts.HeartModule.ErrorHandler;
-using Heart_Module.Data.Scripts.HeartModule.Projectiles.StandardClasses;
-using Sandbox.ModAPI;
+﻿using Heart_Module.Data.Scripts.HeartModule.Projectiles.StandardClasses;
 using System.Collections.Generic;
-using VRage.Utils;
 
 namespace Heart_Module.Data.Scripts.HeartModule.Projectiles
 {
