+++ Orrery Combat Framework - Heart Module/Data/Scripts/HeartModule/Weapons/StandardClasses/WeaponDefinitionBase.cs
@@ -1,4 +1,4 @@
-﻿using Heart_Module.Data.Scripts.HeartModule.Projectiles.StandardClasses;
+﻿using Heart_Module.Data.Scripts.HeartModule.Definitions.StandardClasses;
 using ProtoBuf;
 using System;
 
@@ -130,24 +130,5 @@ namespace Heart_Module.Data.Scripts.HeartModule.Weapons.StandardClasses
 
         public bool HasShootParticle => !ShootParticle?.Equals("") ?? false;
         public bool HasReloadParticle => !ReloadParticle?.Equals("") ?? false;
-    }
-
-    public enum IFF_Enum
-    {
-        None = 0,
-        TargetSelf = 1,
-        TargetEnemies = 2,
-        TargetFriendlies = 4,
-        TargetNeutrals = 8,
-        TargetUnique = 16,
-    }
-
-    public enum TargetType_Enum
-    {
-        None = 0,
-        TargetGrids = 1,
-        TargetProjectiles = 2,
-        TargetCharacters = 4,
-        TargetUnique = 8,
     }
 }
