+++ Orrery Combat Framework - Heart Module/Data/Scripts/HeartModule/Weapons/WeaponDefinitionManager.cs
@@ -1,9 +1,5 @@
-﻿using Heart_Module.Data.Scripts.HeartModule.ErrorHandler;
-using Heart_Module.Data.Scripts.HeartModule.ExceptionHandler;
-using Heart_Module.Data.Scripts.HeartModule.Weapons.StandardClasses;
-using System;
+﻿using Heart_Module.Data.Scripts.HeartModule.Weapons.StandardClasses;
 using System.Collections.Generic;
-using VRage.Utils;
 
 namespace Heart_Module.Data.Scripts.HeartModule.Weapons
 {
